backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.2
numpy==2.2.6
sentence-transformers==5.1.0
scikit-learn==1.7.1
PyMuPDF==1.26.4
python-docx==1.2.0
python-multipart==0.0.20
pydantic==2.11.7
python-jose[cryptography]==3.5.0
passlib==1.7.4
SQLAlchemy==2.0.42
email-validator==2.3.0
